habitflow-ai/habitflow-ai/requirements.txt
# HabitFlow AI - MVP Requirements
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
python-multipart==0.0.6
jinja2==3.1.2

